Fix bug with mkdir race-condition
The original code, when run with MPI (as intended), suffers from a race-
condition where multiple processes attempt to create a new directory at
roughly the same time. As it happens, sometimes the new directory gets
created by another MPI process in between the existence-check and the attempt
by this process to do so, thus triggering a OSError that the directory already
exists.

The simple solution is to put it in a try-except block, which preserves
functionality, but it does mean that other errors will not be caught.

See https://stackoverflow.com/questions/1586648/race-condition-creating-
folder-in-python for more details.